+++ src/js/bitcoinjs-extensions.js
@@ -691,6 +691,28 @@ libs.bitcoin.networks.digitalcoin = {
   wif: 0x9e,
 };
 
+libs.bitcoin.networks.divi = {
+  messagePrefix: '\x19Divi Signed Message:\n',
+  bip32: {
+    public: 0x022d2533,
+    private: 0x0221312b,
+  },
+  pubKeyHash: 0x1e,
+  scriptHash: 0x0d,
+  wif: 0xd4,
+};
+
+libs.bitcoin.networks.divitestnet = {
+  messagePrefix: '\x19Divi Signed Message:\n',
+  bip32: {
+    public: 0x3a8061a0,
+    private: 0x3a805837,
+  },
+  pubKeyHash: 0x8b,
+  scriptHash: 0x13,
+  wif: 0xef,
+};
+
 libs.bitcoin.networks.ecoin = {
   messagePrefix: '\x18eCoin Signed Message:\n',
   bip32: {
@@ -1531,6 +1553,17 @@ libs.bitcoin.networks.bitcoinprivate = {
   wif: 0x80,
 };
 
+libs.bitcoin.networks.bitcoinprivatetestnet = {
+  messagePrefix: '\x18BitcoinPrivate Signed Message:\n',
+  bip32: {
+    public: 0x043587CF,
+    private: 0x04358394,
+  },
+  pubKeyHash: 0x1957,
+  scriptHash: 0x19E0,
+  wif: 0xEF,
+};
+
 libs.bitcoin.networks.bitcoinz = {
   messagePrefix: '\x18BitcoinZ Signed Message:\n',
   bip32: {
@@ -1864,4 +1897,16 @@ libs.bitcoin.networks.argoneum = {
   pubKeyHash: 0x32,
   scriptHash: 0x61,
   wif: 0xbf
+};
+
+libs.bitcoin.networks.particl = {
+  messagePrefix: '\x18Bitcoin Signed Message:\n',
+  bech32: 'pw',
+  bip32: {
+    public: 0x696e82d1,
+    private: 0x8f1daeb8,
+  },
+  pubKeyHash: 0x38,
+  scriptHash: 0x3c,
+  wif: 0x6c,
 };